testonlinetf218

+++ hourglass_tensorflow/handlers/train.py
@@ -89,7 +89,7 @@ class HTFTrainHandler(_HTFTrainHandler):
         self._learning_rate = self._instantiate(self.config.learning_rate)
         self._loss = self._instantiate(self.config.loss)
         self._optimizer = self._instantiate(
-            self.config.optimizer, learning_rate=np.float32(self._learning_rate)
+            self.config.optimizer, learning_rate=self._learning_rate
         )
         self._metrics = [obj.init() for obj in self.config.metrics]
         self._callbacks = [obj.init() for obj in self.config.callbacks]
